requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai>=0.3.0
faiss-cpu>=1.8.0
sentence-transformers>=2.2.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic>=2.5.0
numpy>=1.24.3
httpx>=0.25.0